% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_tmhs.R
\name{are_tmhs}
\alias{are_tmhs}
\title{Are the sequences TMHs?}
\usage{
are_tmhs(protein_sequences)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}
}
\description{
Are the sequences TMHs?
}
\examples{
library(testthat)

sequences <- c(
  "VVIILTIAGNILVIMAVSLE",
  "VVIILTIRGNILVIMAVSLE"
)
expect_equal(c(TRUE, FALSE), are_tmhs(sequences))
}
