% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/are_topology_sequences.R
\name{are_topology_sequences}
\alias{are_topology_sequences}
\title{Are all the one or more topology sequences valid?}
\usage{
are_topology_sequences(topology_sequences)
}
\arguments{
\item{topology_sequences}{one or more topology sequences,
for example \code{c("iiiimmmmooooo", "iiiii")}}
}
\description{
Are all the one or more topology sequences valid?
}
\examples{
# Valid topology sequences
are_topology_sequences("oooooMMMMMiiiii")
are_topology_sequences(c("oooooMMMMMiiiii", "iiiii"))

# Invalid topology sequences
are_topology_sequences("nonsense")
are_topology_sequences(c("oooooMMMMMiiiii", "nonsense"))
are_topology_sequences(c())
are_topology_sequences(NULL)
}
