% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tmhmm_installation.R
\name{check_tmhmm_installation}
\alias{check_tmhmm_installation}
\title{Checks the installation of TMHMM.
Throws a helpful error message if incomplete,
else does nothing}
\usage{
check_tmhmm_installation(folder_name = get_default_tmhmm_folder())
}
\arguments{
\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\value{
Nothing
}
\description{
Checks the installation of TMHMM.
Throws a helpful error message if incomplete,
else does nothing
}
\author{
Richel J.C. Bilderbeek
}
