% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_text_to_tibble.R
\name{fasta_text_to_tibble}
\alias{fasta_text_to_tibble}
\title{Convert FASTA text to a tibble}
\usage{
fasta_text_to_tibble(fasta_text)
}
\arguments{
\item{fasta_text}{the content of a FASTA file as text,
for example \code{c(">Protein 1", "FAMILYVW")}}
}
\description{
Convert FASTA text to a tibble
}
\examples{
topology_text <- c(">Protein A", "iiiimmmmmooooo")
fasta_text_to_tibble(topology_text)
}
