% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_tmhmm_bin_path.R
\name{get_default_tmhmm_bin_path}
\alias{get_default_tmhmm_bin_path}
\title{Get the default path to the TMHMM binary}
\usage{
get_default_tmhmm_bin_path()
}
\value{
the default path to the TMHMM binary
}
\description{
Get the default path to the TMHMM binary
}
\author{
Richèl J.C. Bilderbeek
}
