% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_default_tmhmm_folder.R
\name{get_default_tmhmm_folder}
\alias{get_default_tmhmm_folder}
\title{Get the path to the folder where this package installs
TMHMM by default}
\usage{
get_default_tmhmm_folder()
}
\value{
the path to the folder where this package installs
  TMHMM by default
}
\description{
Get the path to the folder where this package installs
TMHMM by default
}
\examples{
get_default_tmhmm_folder()
}
\author{
Richèl J.C. Bilderbeek
}
