% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tmhmm_url.R
\name{get_tmhmm_url}
\alias{get_tmhmm_url}
\title{Get an URL to download TMHMM from.
These URLs expire.}
\usage{
get_tmhmm_url()
}
\description{
Get an URL to download TMHMM from.
These URLs expire.
}
\author{
Richel J.C. Bilderbeek
}
