% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tmhmm.R
\name{install_tmhmm}
\alias{install_tmhmm}
\title{Install NetMHC2pan to a local folder}
\usage{
install_tmhmm(download_url = NULL,
  folder_name = get_default_tmhmm_folder(), verbose = FALSE)
}
\arguments{
\item{download_url}{the URL to download TMHMM from}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}

\item{verbose}{set to TRUE for more output}
}
\description{
Install NetMHC2pan to a local folder
}
\author{
Richel J.C. Bilderbeek
}
