% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tmhmm.R
\name{install_tmhmmn}
\alias{install_tmhmmn}
\title{Install NetMHC2pan to a local folder}
\usage{
install_tmhmmn(download_url = NULL,
  folder_name = rappdirs::user_data_dir(), verbose = FALSE,
  os = rappdirs::app_dir()$os)
}
\arguments{
\item{verbose}{set to TRUE for more output}
}
\description{
Install NetMHC2pan to a local folder
}
\author{
Richel J.C. Bilderbeek
}
