% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_tmhmm.R
\name{install_tmhmmn_bin}
\alias{install_tmhmmn_bin}
\title{Install the NetMHC2pan binary to a local folder}
\usage{
install_tmhmmn_bin(download_url = get_download_url(),
  folder_name = rappdirs::user_data_dir(), verbose = FALSE,
  os = rappdirs::app_dir()$os)
}
\arguments{
\item{verbose}{set to TRUE for more output}
}
\description{
Install the NetMHC2pan binary to a local folder
}
\author{
Richel J.C. Bilderbeek
}
