% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_installed.R
\name{is_tmhmm_installed}
\alias{is_tmhmm_installed}
\title{Measure if TMHMM is installed locally}
\usage{
is_tmhmm_installed(folder_name = get_default_tmhmm_folder(),
  os = rappdirs::app_dir()$os)
}
\value{
TRUE is TMHMM is installed locally,
  FALSE otherwise
}
\description{
Measure if TMHMM is installed locally
}
\author{
Richel J.C. Bilderbeek
}
