% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_result.R
\name{is_tmhmm_result}
\alias{is_tmhmm_result}
\title{Is the text a locatome text, as would be in a FASTA file?}
\usage{
is_tmhmm_result(tmhmm_result)
}
\description{
Is the text a locatome text, as would be in a FASTA file?
}
