% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_set_up.R
\name{is_tmhmm_set_up}
\alias{is_tmhmm_set_up}
\title{Determine if TMHMM is set up}
\usage{
is_tmhmm_set_up(tmhmm_bin_filename = get_default_tmhmm_bin_path())
}
\value{
TRUE if TMHMM is set up, FALSE otherwise
}
\description{
Determine if TMHMM is set up
}
\author{
Richel J.C. Bilderbeek
}
