% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_topology_sequence.R
\name{is_topology_sequence}
\alias{is_topology_sequence}
\title{Is the topology sequence valid?}
\usage{
is_topology_sequence(topology_sequence)
}
\arguments{
\item{topology_sequence}{a topology sequence,
for example \code{"iiiimmmmmoooo"}}
}
\description{
Is the topology sequence valid?
}
\examples{
# A valid topology sequence
is_topology_sequence("oooooMMMMMiiiii")

# An invalid topology sequence
is_topology_sequence("nonsense")
}
