% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_topology_text.R
\name{is_topology_text}
\alias{is_topology_text}
\title{Is the text a topology text?}
\usage{
is_topology_text(topology_text)
}
\arguments{
\item{topology_text}{the topology as text,
that alternates protein names and topology sequences.
This is the same form as a FASTA file,
for example \code{c(">Protein A", "iiiimmmmmoooo")}}
}
\value{
TRUE if the text is a topology text
}
\description{
Is the text a topology text?
}
\examples{
topology_text <- c(
  ">protein name",
  "ooooMMMMiiiii"
)
is_topology_text(topology_text)
}
