% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_url_valid.R
\name{is_url_valid}
\alias{is_url_valid}
\title{Is the download URL valid?}
\usage{
is_url_valid(download_url = get_tmhmm_url())
}
\arguments{
\item{download_url}{the URL to download TMHMM from}
}
\description{
The download link expires after 4 hours.
}
