% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatome_to_df.R
\name{locatome_to_df}
\alias{locatome_to_df}
\title{Convert a multi-line locatome to a tidy data frame}
\usage{
locatome_to_df(locatome)
}
\description{
Convert a multi-line locatome to a tidy data frame
}
