% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_predict_topology_from_sequence.R
\name{mock_predict_topology_from_sequence}
\alias{mock_predict_topology_from_sequence}
\title{Mock the topology prediction.}
\usage{
mock_predict_topology_from_sequence(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\value{
a topology sequence,
  for example \code{"iiiimmmmmoooo"}
}
\description{
Use \link{predict_topology_from_sequence} for a real prediction.
}
\examples{
protein_sequence <- paste0(
  "QEKNWSALLTAVVIILTIAGNILVIMAVSLEKKLQNATNYFLM",
  "SLAIADMLLGFLVMPVSMLTILYGYRWP"
)
mock_predict_topology_from_sequence(protein_sequence)
}
\author{
Richèl J.C. Bilderbeek
}
