% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_run_tmhmm_to_file.R
\name{mock_run_tmhmm_to_file}
\alias{mock_run_tmhmm_to_file}
\title{Mock \link{run_tmhmm_to_file}}
\usage{
mock_run_tmhmm_to_file(
  fasta_filename,
  tmhmm_filename,
  one_protein_fasta_filename = "irrelevant",
  folder_name = "irrelevant"
)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{tmhmm_filename}{filename to write the TMHMM results to}

\item{one_protein_fasta_filename}{name of a FASTA file,
that contains one name and one sequence}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\description{
Mock \link{run_tmhmm_to_file}
}
\examples{
# Write results to a temporary file
tmhmm_filename <- tempfile()
run_tmhmm_to_file(
  fasta_filename = system.file(
    "extdata", "tmhmm.fasta", package = "tmhmm"
  ),
  tmhmm_filename = tmhmm_filename
)

# Result is written to 'tmhmm_filename', cleaning up
file.remove(tmhmm_filename)
}
