% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_fasta_file.R
\name{parse_fasta_file}
\alias{parse_fasta_file}
\title{Parse a FASTA file to a table
with a \code{name} and \code{sequence} column}
\usage{
parse_fasta_file(fasta_filename)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}
}
\value{
a \link[tibble]{tibble}
  with a \code{name} and \code{sequence} column
}
\description{
Parse a FASTA file to a table
with a \code{name} and \code{sequence} column
}
