% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_locatome.R
\name{plot_locatome}
\alias{plot_locatome}
\title{Plot the topology from text}
\usage{
plot_locatome(locatome)
}
\arguments{
\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}
}
\description{
Plot the topology from text
}
\examples{
topology_text <- c(">Protein A", "iiiimmmmmooooo")
plot_topology_text(topology_text)
}
