% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_topology_text.R
\name{plot_topology_text}
\alias{plot_topology_text}
\title{Plot the topology from text}
\usage{
plot_topology_text(topology_text)
}
\arguments{
\item{topology_text}{the topology as text,
in a similar form as a FASTA file,
for example \code{c(">Protein A", "iiiimmmmmoooo")}}
}
\description{
Plot the topology from text
}
\examples{
topology_text <- c(">Protein A", "iiiimmmmmooooo")
plot_topology_text(topology_text)
}
