% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_topologies_from_sequences.R
\name{predict_topologies_from_sequences}
\alias{predict_topologies_from_sequences}
\title{Run TMHMM directy on a protein sequence}
\usage{
predict_topologies_from_sequences(
  protein_sequences,
  folder_name = get_default_tmhmm_folder(),
  temp_fasta_filename = tempfile(fileext = ".fasta"),
  temp_topology_filename = tempfile(fileext = ".topo")
)
}
\arguments{
\item{protein_sequences}{one or more protein sequences}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}

\item{temp_fasta_filename}{filename to temporarily
save the protein names and sequences.
This file will be deleted when this
function completed successfully}

\item{temp_topology_filename}{filename to temporarily
save the topology. This file will be deleted when this
function completed successfully}
}
\value{
one or more topology sequences,
  for example \code{c("iiiimmmmooooo", "iiiii")}
}
\description{
Run TMHMM directy on a protein sequence
}
\examples{
if (is_tmhmm_installed()) {
  protein_sequence <- paste0(
    "QEKNWSALLTAVVIILTIAGNILVIMAVSLEKKLQNATNYFLM",
    "SLAIADMLLGFLVMPVSMLTILYGYRWP"
  )
  predict_topology_from_sequence(protein_sequence)
}
}
\seealso{
use \link{mock_predict_topologies_from_sequences}
to mock the prediction of protein sequences, as can be useful
in testing
}
\author{
Richèl J.C. Bilderbeek
}
