% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_topology_from_sequence.R
\name{predict_topology_from_sequence}
\alias{predict_topology_from_sequence}
\title{Run TMHMM directy on a protein sequence}
\usage{
predict_topology_from_sequence(
  protein_sequence,
  folder_name = get_default_tmhmm_folder(),
  temp_fasta_filename = tempfile(fileext = ".fasta"),
  temp_topology_filename = tempfile(fileext = ".topo")
)
}
\arguments{
\item{protein_sequence}{a protein sequence, with
the amino acids as capitals, for
example \code{MEILCEDNTSLSSIPNSL}}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}

\item{temp_fasta_filename}{filename to temporarily
save the protein names and sequences.
This file will be deleted when this
function completed successfully}

\item{temp_topology_filename}{filename to temporarily
save the topology. This file will be deleted when this
function completed successfully}
}
\value{
a topology sequence,
  for example \code{"iiiimmmmmoooo"}
}
\description{
Run TMHMM directy on a protein sequence
}
\examples{
if (is_tmhmm_installed()) {
  protein_sequence <- paste0(
    "QEKNWSALLTAVVIILTIAGNILVIMAVSLEKKLQNATNYFLM",
    "SLAIADMLLGFLVMPVSMLTILYGYRWP"
  )
  predict_topology_from_sequence(protein_sequence)
}
}
\author{
Richèl J.C. Bilderbeek
}
