% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tmhmm.R
\name{run_tmhmm}
\alias{run_tmhmm}
\title{Run TMHMM}
\usage{
run_tmhmm(fasta_filename)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}
}
\value{
a data frame
}
\description{
Run TMHMM
}
\author{
Richel J.C. Bilderbeek
}
