% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tmhmm_on_one_protein_fasta_file.R
\name{run_tmhmm_on_one_protein_fasta_file}
\alias{run_tmhmm_on_one_protein_fasta_file}
\title{Run TMHMM on a FASTA file with one protein}
\usage{
run_tmhmm_on_one_protein_fasta_file(
  one_protein_fasta_filename,
  folder_name = get_default_tmhmm_folder()
)
}
\arguments{
\item{one_protein_fasta_filename}{name of a FASTA file,
that contains one name and one sequence}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\value{
text similar to a FASTA file, yet with the
  predicted topology instead of the protein sequence
}
\description{
This is the function that actually calls TMHMM.
The usage of TMHMM us split up per protein, as R cannot
handle communication via \code{stdin} that is too long,
hence, bigger FASTA files get truncated, and a topology of
less proteins is returned.
}
\details{
Use \link{run_tmhmm} to call TMHMM with a FASTA file with multiple lines.
}
\examples{
if (is_tmhmm_installed()) {
  one_protein_fasta_filename <- system.file(
    "extdata", "tmhmm.fasta", package = "tmhmm"
  )
  topology_text <- run_tmhmm(one_protein_fasta_filename)
  message(topology_text, sep = "\n")
}
}
\author{
Richèl J.C. Bilderbeek
}
