% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_tibble_as_fasta_file.R
\name{save_tibble_as_fasta_file}
\alias{save_tibble_as_fasta_file}
\title{Save the first two columns of a tibble as a FASTA file}
\usage{
save_tibble_as_fasta_file(t, fasta_filename)
}
\description{
Save the first two columns of a tibble as a FASTA file
}
