% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_tmhmm.R
\name{set_up_tmhmm}
\alias{set_up_tmhmm}
\title{Set up TMHMM}
\usage{
set_up_tmhmm(tmhmm_bin_filename = get_default_tmhmm_bin_path())
}
\arguments{
\item{tmhmm_bin_filename}{path to the TMHMM binary file.
Its default path can be ontained by using
\link{get_default_tmhmm_bin_path}}
}
\description{
Set up TMHMM
}
\author{
Richel J.C. Bilderbeek
}
