% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_locatome.R
\name{tally_locatome}
\alias{tally_locatome}
\title{Tally each location in a locatome}
\usage{
tally_locatome(locatome)
}
\arguments{
\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}
}
\description{
Tally each location in a locatome
}
\examples{
tally_locatome("iMMooo")
}
\author{
Richèl J.C. Bilderbeek
}
