% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_tmhs.R
\name{tally_tmhs}
\alias{tally_tmhs}
\title{Count the number of transmembrane helices in a topology}
\usage{
tally_tmhs(topology)
}
\value{
a \link[tibble]{tibble} with the number of
TMHs per protein
}
\description{
Count the number of transmembrane helices in a topology
}
\examples{
if (is_tmhmm_installed()) {
  tally_tmhs(
    predict_topology(
      fasta_filename <- system.file(
        "extdata", "example_2.fasta", package = "tmhmm"
      )
    )
  )
}
}
