% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmhmm_self_test.R
\name{tmhmm_self_test}
\alias{tmhmm_self_test}
\title{Self-test the TMHMM installation}
\usage{
tmhmm_self_test()
}
\description{
Self-test the TMHMM installation
}
\examples{
if (is_tmhmm_installed()) {
  tmhmm_self_test())
}
}
\author{
Richèl J.C. Bilderbeek
}
