% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_tmhmm.R
\name{uninstall_tmhmm}
\alias{uninstall_tmhmm}
\title{Uninstall TMHMM}
\usage{
uninstall_tmhmm(folder_name = get_default_tmhmm_folder())
}
\arguments{
\item{folder_name}{name of the folder
where the TMHMM files are installed.
The name of the TMHMM binary file will be at
\code{[folder_name]/tmhmm-2.0c/bin/tmhmm}}
}
\description{
Uninstall TMHMM
}
\author{
Richèl J.C. Bilderbeek
}
