% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_topology.R
\name{check_topology}
\alias{check_topology}
\title{Check if the topology is valid.}
\usage{
check_topology(topology)
}
\arguments{
\item{topology}{a \link[tibble]{tibble} with two columns,
named \code{name} for a protein's name and \code{topology},
for its predicted topology (e.g. \code{ooooMMMMiiiii}).}
}
\value{
Nothing. Will \link{stop} with a helpful error message if
the topology is invalid.
}
\description{
Check if the argument is of the same type as a predicted topology,
as can be created with \link{run_tmhmm}.
Will \link{stop} if not.
}
\author{
Richèl J.C. Bilderbeek
}
