% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  download_url,
  fasta_filename,
  folder_name,
  locatome,
  protein_sequence,
  protein_sequences,
  tmhmm_bin_filename,
  tmhmm_filename,
  verbose
)
}
\arguments{
\item{download_url}{the URL to download TMHMM from}

\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}

\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{tmhmm_bin_filename}{path to the TMHMM binary file.
Its default path can be ontained by using
\link{get_default_tmhmm_bin_path}}

\item{tmhmm_filename}{filename to write the TMHMM results to}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
