% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_params_doc.R
\name{default_params_doc}
\alias{default_params_doc}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
default_params_doc(
  download_url,
  fasta_filename,
  fasta_text,
  folder_name,
  locatome,
  one_protein_fasta_filename,
  protein_sequence,
  protein_sequences,
  temp_fasta_filename,
  temp_topology_filename,
  tmhmm_bin_filename,
  tmhmm_filename,
  tmhmm_result,
  topology,
  topology_sequence,
  topology_sequences,
  topology_text,
  verbose
)
}
\arguments{
\item{download_url}{the URL to download TMHMM from}

\item{fasta_filename}{path to a FASTA file}

\item{fasta_text}{the content of a FASTA file as text,
for example \code{c(">Protein 1", "FAMILYVW")}}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}

\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}

\item{one_protein_fasta_filename}{name of a FASTA file,
that contains one name and one sequence}

\item{protein_sequence}{a protein sequence}

\item{protein_sequences}{one or more protein sequences}

\item{temp_fasta_filename}{filename to temporarily
save the protein names and sequences.
This file will be deleted when this
function completed successfully}

\item{temp_topology_filename}{filename to temporarily
save the topology. This file will be deleted when this
function completed successfully}

\item{tmhmm_bin_filename}{path to the TMHMM binary file.
Its default path can be ontained by using
\link{get_default_tmhmm_bin_path}}

\item{tmhmm_filename}{filename to write the TMHMM results to}

\item{tmhmm_result}{the results of a call to TMHMM,
for example, \code{c(">protein name", "ooooMMMMiiiii")}}

\item{topology}{a \link[tibble]{tibble} with two columns,
named \code{name} for a protein's name and \code{topology},
for its predicted topology (e.g. \code{ooooMMMMiiiii}).}

\item{topology_sequence}{a topology sequence,
for example \code{"iiiimmmmmoooo"}}

\item{topology_sequences}{one or more topology sequences,
for example \code{c("iiiimmmmooooo", "iiiii")}}

\item{topology_text}{the topology as text,
that alternates protein names and topology sequences.
This is the same form as a FASTA file,
for example \code{c(">Protein A", "iiiimmmmmoooo")}}

\item{verbose}{set to TRUE for more output}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
\note{
This is an internal function, so it should be marked with
  \code{@noRd}. This is not done, as this will disallow all
  functions to find the documentation parameters
}
\author{
Richèl J.C. Bilderbeek
}
