% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_locatome_line.R
\name{is_locatome_line}
\alias{is_locatome_line}
\title{Is the line of text the locatome, as used within
a FASTA filename?}
\usage{
is_locatome_line(line)
}
\arguments{
\item{line}{line of text from a FASTA filename}
}
\description{
Is the line of text the locatome, as used within
a FASTA filename?
}
\examples{
# A valid locatome line
is_locatome_line("oooooMMMMMiiiii")

# An invalid locatome line
is_locatome_line("nonsense"))
}
