% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_protein_name_line.R
\name{is_protein_name_line}
\alias{is_protein_name_line}
\title{Is the line of text the name of a protein, as used within
a FASTA filename?}
\usage{
is_protein_name_line(line)
}
\arguments{
\item{line}{line of text from a FASTA filename}
}
\description{
Is the line of text the name of a protein, as used within
a FASTA filename?
}
\examples{
library(testthat)

expect_true(is_protein_name_line(">5H2A_CRIGR"))
expect_false(is_protein_name_line("5H2A_CRIGR"))
expect_false(is_protein_name_line("oooooMMMMMiiiii"))
expect_false(is_protein_name_line(NA))
expect_false(is_protein_name_line(NULL))
expect_false(is_protein_name_line(""))
}
