% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmh.R
\name{is_tmh}
\alias{is_tmh}
\title{Determine if the protein sequence contains at least one TMH.}
\usage{
is_tmh(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence}
}
\description{
Determine if the protein sequence contains at least one TMH.
}
\examples{
# A TMH
is_tmh("VVIILTIAGNILVIMAVSLE")

# A protein without a TMH
is_tmh("VVIILTIRGNILVIMAVSLE")
}
