% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_installed.R
\name{is_tmhmm_bin_installed}
\alias{is_tmhmm_bin_installed}
\title{Measure if TMHMM binary is installed locally}
\usage{
is_tmhmm_bin_installed(folder_name = get_default_tmhmm_folder())
}
\arguments{
\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\value{
TRUE is TMHMM binary is installed locally,
  FALSE otherwise
}
\description{
Measure if TMHMM binary is installed locally
}
\author{
Richel J.C. Bilderbeek
}
