% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_result.R
\name{is_tmhmm_result}
\alias{is_tmhmm_result}
\title{Is the text a TMHMM result text?}
\usage{
is_tmhmm_result(tmhmm_result)
}
\arguments{
\item{tmhmm_result}{the results of a call to TMHMM,
for example, \code{c(">protein name", "ooooMMMMiiiii")}}
}
\value{
TRUE if the text is a TMHMM result text
}
\description{
Is the text a TMHMM result text?
}
\examples{
tmhmm_result <- c(
  ">protein name",
  "ooooMMMMiiiii"
)
is_topology_text(topology_text = tmhmm_result)
}
