% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_tmhmm_set_up.R
\name{is_tmhmm_set_up}
\alias{is_tmhmm_set_up}
\title{Determine if TMHMM is set up}
\usage{
is_tmhmm_set_up(folder_name = get_default_tmhmm_folder())
}
\arguments{
\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\value{
TRUE if TMHMM is set up, FALSE otherwise
}
\description{
Determine if TMHMM is set up
}
\author{
Richèl J.C. Bilderbeek
}
