% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locatome_to_df.R
\name{locatome_to_df}
\alias{locatome_to_df}
\title{Convert a multi-line locatome to a tidy data frame}
\usage{
locatome_to_df(topology_text, locatome = character(0))
}
\arguments{
\item{topology_text}{the topology as text,
that alternates protein names and topology sequences.
This is the same form as a FASTA file,
for example \code{c(">Protein A", "iiiimmmmmoooo")}}

\item{locatome}{text similar to a FASTA file, with protein names
and the locations of its amino acids}
}
\description{
Convert a multi-line locatome to a tidy data frame
}
\examples{
topology_text <- c(">Protein A", "iiiimmmmmooooo")
fasta_text_to_tibble(topology_text)
}
