% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_run_tmhmm.R
\name{mock_run_tmhmm}
\alias{mock_run_tmhmm}
\title{Mock \link{run_tmhmm}}
\usage{
mock_run_tmhmm(fasta_filename)
}
\arguments{
\item{fasta_filename}{path to a FASTA file}
}
\value{
the topology as text, in a similar form as a FASTA file,
  for example \code{c(">Protein A", "iiiimmmmmoooo")}.
  Use \link{predict_topology} to get the same result
  in the form of a \link[tibble]{tibble}
}
\description{
Mock \link{run_tmhmm}
}
\examples{
fasta_filename <- system.file("extdata", "tmhmm.fasta", package = "tmhmm")
topology_text <- mock_run_tmhmm(fasta_filename)
cat(topology_text, sep = "\n")
}
\author{
Richèl J.C. Bilderbeek
}
