% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tmhmm.R
\name{run_tmhmm}
\alias{run_tmhmm}
\title{Run TMHMM}
\usage{
run_tmhmm(fasta_filename, folder_name = get_default_tmhmm_folder())
}
\arguments{
\item{fasta_filename}{path to a FASTA file}

\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\value{
a character vector with the locatome
}
\description{
Run TMHMM
}
\examples{
  testit::assert(is_tmhmm_installed())
  fasta_filename <- system.file("extdata", "tmhmm.fasta", package = "tmhmm")
  locatome <- run_tmhmm(fasta_filename)
  cat(locatome, sep = "\\n")
}
\author{
Richel J.C. Bilderbeek
}
