% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_tmhmm_on_sequence.R
\name{run_tmhmm_on_sequence}
\alias{run_tmhmm_on_sequence}
\title{Run TMHMM directy on a protein sequence}
\usage{
run_tmhmm_on_sequence(protein_sequence)
}
\arguments{
\item{protein_sequence}{a protein sequence, with
the amino acids as capitals, for
example `MEILCEDNTSLSSIPNSL`}
}
\value{
the locatome
}
\description{
Run TMHMM directy on a protein sequence
}
\examples{
protein_sequence <- paste0(
  "QEKNWSALLTAVVIILTIAGNILVIMAVSLEKKLQNATNYFLM",
  "SLAIADMLLGFLVMPVSMLTILYGYRWP"
)
locatome <- run_tmhmm_on_sequence(protein_sequence)
expected_locatome <- paste0(
  "iiiiiiMMMMMMMMMMMMMMMMMMMMMMMoooooooooooooo",
  "MMMMMMMMMMMMMMMMMMMMMMMiiiii"
)
expect_equal(locatome, expected_locatome)
}
