% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up_tmhmm.R
\name{set_up_tmhmm}
\alias{set_up_tmhmm}
\title{Set up TMHMM}
\usage{
set_up_tmhmm(folder_name = get_default_tmhmm_folder())
}
\arguments{
\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\description{
Set up TMHMM
}
\author{
Richèl J.C. Bilderbeek
}
