% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally_tmhs.R
\name{tally_tmhs}
\alias{tally_tmhs}
\title{Count the number of transmembrane helices in a topology}
\usage{
tally_tmhs(topology)
}
\value{
a \link[tibble]{tibble} with the number of
TMHs per protein
}
\description{
Count the number of transmembrane helices in a topology
}
\examples{
if (is_pureseqtm_installed()) {
  tally_tmhs(
    predict_topology(
      get_example_filename("1bhaA.fasta")
    )
  )
}
}
