% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uninstall_tmhmm.R
\name{uninstall_tmhmm}
\alias{uninstall_tmhmm}
\title{Uninstall TMHMM}
\usage{
uninstall_tmhmm(folder_name = get_default_tmhmm_folder())
}
\arguments{
\item{folder_name}{superfolder of TMHMM.
The superfolder's name is \code{/home/[user_name]/.local/share}
by default, as can be obtained by
\link{get_default_tmhmm_folder}}
}
\value{
Nothing
}
\description{
Uninstall TMHMM. Will stop if TMHMM is not installed.
}
\examples{
\donttest{
  if (is_url_valid() && is_tmhmm_installed()) {
    uninstall_tmhmm()
    install_tmhmm()
  }
}
}
\author{
Richèl J.C. Bilderbeek
}
